- fixed some uninitialized stuff in particle rendering. - added gl_ClampLight to all places where the light level is set.
Incorporates small part of c47c7421a36f4594f1637367607ba09666c44b35

# Conflicts:
#	src/gl/scene/gl_sprite.cpp

+++ src/gl/scene/gl_wall.h
@@ -119,7 +119,7 @@ public:
 	
 	float ViewDistance;
 
-	BYTE lightlevel;
+	int lightlevel;
 	BYTE type;
 	BYTE flags;
 	short rellight;
